Update weather_requester to handle locations not found
When the 3rd party api can not find a location it does not seem to
return an 'error' in the manner we were expecting previously.

Now we check a 'cod' key in the response body and if it has a value of
404 then we know the backend api could not find / match the requested
location and we should inform the requester.

This stops the application from crashing out if the location could not
be found.